requirements.txt
biopython==1.75
ujson==1.35